requirements.txt
aioredis~=1.3.1
click
bmt==1.4.3
deepdiff==8.6.1
fastapi~=0.108.0
httptools
jsonschema~=4.6.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-asyncio==0.18.3
pyyaml~=6.0
reasoner-pydantic==4.1.5
redis~=3.5.3
redis-py-cluster==2.1.3
requests==2.32.4
testcontainers==3.6.1
uvicorn
uvloop
gunicorn==23.0.0
orjson==3.9.15
httpx==0.23.0

# To support Open Telemetry
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
